backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-multipart==0.0.20
pandas==2.3.1
openpyxl==3.1.5
rapidfuzz==3.13.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.11.7
python-dotenv==1.1.1
alembic==1.16.4
